From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: fireball spell while engulfed
Bulletproof the "bouncy" version of the fireball spell (I never did get a
crash, but verified an out-of-bounds error).  Based on my reading of the
code, this could fail even if not engulfed, as long as you're near the edge
of the screen.  When engulfed, an additional u.uswallow check is needed to
match an initial check done in throwspell().

I also made the block comment at the top of the switch statement more
accurate for the current spell system.

Modified Files:
	doc/fixes33.1 src/spell.c