From: Mark W Modrall <modrall@world.std.com> Subject: fix for mac port "Hungry" bug
        finally managed to reproduce the "Hungry" display bug (where
someone would eat something but the hunger status wouldn't update on
the screen...  found a mail message i hadn't seen before in the unresolved
lists last friday from one of the people who had reported it to the effect
that they were using ? at the eat prompt to bring up a food menu (something
i'm not in the habit of doing); that turned out to be the way to reproduce
it.
        it turned out that the problem happens when the food menu overlaps
the map window.  when you select some food and close the menu, it issues
an update message for the screen.  Before you get back around to it in the
event loop, bite() has posted the screen change, but it defers rendering
until the next display_nhwindow() call.

        when the os update screen message gets processed, the os only draws
that part from the overlapped windows, even though the code redraws the
whole screen.  So the internal state of the code thought it had redrawn the
hunger status, but the os had clipped the rendering to only the overlap
region.

        fixed by adding routines to store the pending invalid region from
the rendering code and to restore it after an Os update message...

	plus some changes to clean up compiler warnings...

Modified Files:
	include/mactty.h include/macwin.h sys/mac/dprintf.c
	sys/mac/maccurs.c sys/mac/macerrs.c sys/mac/macfile.c
	sys/mac/macsnd.c sys/mac/mactty.c sys/mac/macunix.c
	sys/mac/macwin.c sys/mac/mgetline.c