From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: player movement rate
     I still think a thorough review and update for movement rates is
needed, but this is a quick fix which sets all the player character
monsters to NORMAL_SPEED.  It does the same for all the quest leaders
and nemeses, although since they're not all human some of these values
might not be appropriate.  It also boosts the soldiers and town watch
from 4 to 10 (which is still a bit less that normal human speed, 12).
That hasn't been play tested; the garrison at Ft.Ludios or in barracks
rooms might swarm a lot nastier now.

Modified Files:
	doc/fixes33.1 src/monst.c