From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: mirrors and unseen monsters
Here's the fix for the reported problem that mirrors show the location of
unseen monsters.  I was thinking of addressing Pat's "TODO" mentioned in the
context diff as well, but that seems like it would require some more
extensive work (since the flash should still affect the unseen monster, but
it shouldn't stop in that case) which I wasn't up to.  Also updated the
block comment on bhit() to make the meaning of INVIS_BEAM more obvious.

Modified Files:
	doc/fixes33.1 src/zap.c