From: michael.allison@mnr.gov.on.ca Subject: patch: not all hats are leather
On the newsgroup, it was reported that damage to helmets such as
a cornutham were getting messages about "Your leather helmet...", even
though they are not made of leather.  It turns out that leather was
hard-coded, likely leftover from a time when all non-metal helmets
were made of leather.

Modified Files:
	include/decl.h src/decl.c src/trap.c