From: Dean Luick <dean@ni.com> Subject: patch: moveloop
This patch fixes up the moveloop to account for the player not accumulating
enough movement points during a turn.  This should fix the "time is
wrong when player is burdened" but reports.

This patch also:
o Makes moves (the turn counter) an absolute time count.
o Makes moves the same as monstermoves (which we will eventually remove).
o Moves a lot of turn code (130+ lines) _inside_ the new turn conditional.
  These are now all based on abolute time.

Things incomplete/possibly wrong/whatever:
o I removed a call to defered_goto near the top of the moveloop.
  I have not confinced myself that that wasn't (seemingly) necessary
  for some obscure reason.
o I moved the incremcent of multi (when < 0) _inside_ the once-per-turn
  area -- thus all no-moves, etc. are all based on absolute time, not
  player movement rates like previously.  This is in contrast to the
  decrement of mul (when > 0) still in the once-per-player-move turn.
  Aside from the inconsistent use of multi (what else is new), this seems
  correct since occupations/stuns/whatever should be expressed in terms
  of absolute time, while multiple moves should be expressed in, well,
  times-to-move.  Its quite possible that there are crossover conditions
  that invalidate this idea.

Modified Files:
	doc/fixes33.1 src/allmain.c