From: michael.allison@mnr.gov.on.ca Subject: patch: control-X
Make a limited version of ^X command available to
non-debug and non-explorer mode players.  Its just a
courtesy function to list name, role, race, gender,
alignment, and deities in one place.  It doesn't reveal
anything that they don't already know, so it shouldn't
be controversial at all.

We had at least one request for something like this
since 3.3.0 was released, and the control-X command
was already reserved for attribute display in
debug mode, and unavailable to regular mode,
so it seemed like a logical place to put it.

Modified Files:
	dat/cmdhelp dat/wizhelp doc/Guidebook.mn doc/Guidebook.tex
	src/cmd.c