From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: stone to flesh should not create genocided monster
As reported, add checks so that if you cast a stone to flesh spell on a
stone golem it will not create a flesh golem when flesh golems have been
genocided.  I actually added a check to newcham() for the case where one is
explicitly changing a monster into a genocided species (such checks already
existed for the random case).  After looking at the callers of newcham(),
the only places that currently can end up with attempting to change into a
genocided monster are, in fact, those that are dealing with golems.

Modified Files:
	doc/fixes33.1 src/mon.c src/zap.c