From: michael.allison@mnr.gov.on.ca Subject: patch:
Removal of saddles from a creature was an outstanding item to be
done prior to 3.3.1.

This allows that.
Side-effect: It causes looting to prompt for a direction. You
can choose > or . to loot a container at your location.
Unless changed, you still have to be on top of a container to
loot it. (Removal of that restriction to allow looting on
an adjacent square is trivial after this patch if desired)

Modified Files:
	src/pickup.c