From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: lava bits
     Fix the two message quibbles pointed out by the beta testers
and fix the burning of inventory when in lava.  (It still doesn't
boil potions though.  Shouldn't it?)  I added amulet hanlding too,
but since none of them are classified as being made out of organic
material, that ends up not having any effect.  Also, allow wishing
for lava just like for pools and thrones and such.

Modified Files:
	doc/fixes33.1 src/objnam.c src/read.c src/trap.c