From: Kevin Hugo <kmhugo@students.wisc.edu> Subject: patch: Ranger quest
This patch implements the Ranger quest levels as I originally
conceived them, rather than ripping off the old Elf quest.  The
start level is a traditional labyrinth, with walls made of trees
and iron bars.  The fill-a levels are arboreal.  The locate level
is based on the old "hunt the wumpus" game (including the dodecahedral
maze, the giant bats, etc).  It's also the first actual use of
trees and iron bars.  I had to reorder the floor types to get
trees to work properly.

Still to do:
- Make quest.dat text more relevant to the new levels.
- Trees are counted in IS_ROCK(), which causes some inappropriate
  messages.

Modified Files:
	dat/Ranger.des include/rm.h