From: Dave Cohrs <cohrs@Legato.COM> Subject: patch: the monster "it" problem
A cause of at least some cases of monsters being fully described instead of
calling them "it" is that the global 'killer' is set in various places when
player death is not imminent.  If x_monnam() is called before the player
dies (and in dbridge.c, it was being set even in cases where a monster, not
a player was dying, causing monsters to be fully described from then on),
the monster gets fully named because x_monnam() assumes that killer is only
set when the player is dead.

I added a new global, delayed_killer, that is now set in cases where death
may happen eventually, and code to move this value to killer when
appropriate, and reset delayed_killer if the sitation corrects itself (yes,
I could have added several globals or enhanced timers for each possible
delayed killer, but that seemed like overkill).  Also added bullet proofing
to ensure that in some cases where killer is set just in case, it gets
reset if the player is going to survive, just like done() does.

The upside of this solution, over one proposed by Christian Bressler, is
that in the drawbridge case, there will now be more specific killer
messages in some cases than you currently see.

Modified Files:
	doc/fixes33.1 include/decl.h src/dbridge.c src/decl.c
	src/dokick.c src/eat.c src/explode.c src/polyself.c src/pray.c
	src/read.c src/timeout.c