From: Janet Walz <jwalz@radix.net> Subject: some eating bugs
Have doeat() set up everything that it will later, after an interruption,
assume is set up, even if the food is rotten to start with.  Using
leftover reqtime and nmod had the potential to cause some of the weird
problems reported, but not all.

Minor nutrition cleanup includes allowing for the possibility of oeaten
being >>'d to 0.

When a rider revives, note that we no longer have a corpse to eat.

Modified Files:
	src/eat.c