From: Janet Walz <jwalz@radix.net> Subject: vegetarianism, from Warwick
This is Warwick's prior patch for vegan/vegetarian stuff, updated
to the current code.  (I had to fiddle with eat.c to accommodate my
recent changes, and uhitm.c needed to be changed to compile.)

This does not reflect the final state of things, but Warwick wanted this
in as a base for future changes.

Modified Files:
	include/extern.h include/mondata.h include/you.h src/cmd.c
	src/dog.c src/eat.c src/muse.c src/objnam.c src/potion.c
	src/uhitm.c