From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: multi-shot fix
     Fix the reported problem of never throwing more than one dart
at a time regardless of skill.  I didn't try to figure out why
objects[] is using negative values for missiles that don't have a
launcher.

     When testing the monster shooting part of this patch, I seem
to have found a different bug:  monsters don't bother wielding bows
any more.  At least the hill orc who picked up the orcish bow and
arrows I dropped never shot more than one at a time--despite the new
code which should enable that--and when it closed to hand-to-hand
range it hit me with a scimitar without any messages about wielding
that.  In other words, it was already wielding the sword at the time
it was throwing arrows at me.  (I'm not going to try to fix that.)

Modified Files:
	doc/fixes33.1 src/dothrow.c src/mthrowu.c