From: michael.allison@mnr.gov.on.ca Subject: patch: misleading error message
The first error if the playground directory is somehow messed up
or incomplete is a failure to open the dungeon file.  We had a couple
of reports of that in the past several months, but with DLB defined
the message is worded funny, and the actual missing file isn't even
mentioned:
 Program initialization has failed.
 Cannot open dungeon description file "dungeon"!

This patch makes it this if DLB is defined:
Program initialization has failed.
Cannot open dungeon description - "dungeon" from "nhdat" file!

Otherwise, if DLB is not defined:
Program initialization has failed.
Cannot open dungeon description - "dungeon" file!

Modified Files:
	src/dungeon.c