From: Janet Walz <jwalz@radix.net> Subject: quest altars
Following a couple trial implementations by Michael and comments by Dave,
a fix for the problem of the neutral Caveman getting a forbidding feeling
at Home (lawful altar as Cavemen used to be purely lawful).

This provides the new keywords coaligned and noncoaligned for altars
(and monsters/priests).  Quest start levels get coaligned altars if
their roles have multiple alignments, and goal levels get noncoaligned
altars then.  The one affected locate level (Arch.des) can fortunately
just get a third type of altar, so we don't have to worry about flipping
altar types for bones levels.

The remaining weirdness is for Wizards, who used to be purely neutral
and are now neutral or chaotic.  Therefore they shouldn't have a chaotic
goal altar.  But this leads to the possibility that the Dark One should
fight for law and the Wizard of the Balance for chaos. :-(  Any better
solutions here?

Modified Files:
	dat/Arch.des dat/Barb.des dat/Caveman.des dat/Wizard.des
	include/align.h src/sp_lev.c util/lev_comp.l