From: Janet Walz <jwalz@radix.net> Subject: stack killers
This started out as an attempt to cut down on the stack size used in the
explode.c/mon.c recursion, which was using massively more space on the
Amiga than normal play.  Most of that stack was going for 2*256 bytes of
buffer per recursion round, when at most one of those buffers was needed.

Instead of the prior thought of creating a new static buffer in explode(),
I found an existing static buffer in eat.c that I've pulled out as the
global killer_buf[], and used in explode.c and mon.c.  (There seem to
be no more static buffers for killers, although there are probably more
on the stack.)

Noting that eat.c was using the system BUFSIZ instead of the NetHack
BUFSZ, I checked the rest of the source and changed some more that
didn't have to do with optimizing I/O.

Modified Files:
	include/decl.h src/decl.c src/eat.c src/explode.c src/mhitu.c
	src/mon.c src/uhitm.c src/zap.c win/X11/winmap.c
	win/X11/wintext.c