From: arromdee@rahul.net Subject: patch: my oops
Duuuuh.

I don't know why I didn't catch this the first time, but my recent fortune
cookie patch (the same one that I forgot to send a fixes33.1 band for) had a
bug in it: I was closing the file inside the loop and initializing the
string outside the loop, which would cause crashes if it ever actually found
a fortune containing "fortune" and tried to skip it.

Modified Files:
	src/rumors.c