From: arromdee@rahul.net Subject: patch: pick-axes
1) Monsters won't try to wield a dwarvish mattock for digging if they are
wearing a shield.
2) Pets will check for wearing objects when picking them up, so pets now will
wear armor, not just keep whatever they already have.
3) Pets will now use pick-axes.  (This still leaves the problem where the
pet won't retain its weapon, however.)
4) The "fix" made a long time ago which checked appr and prevented the monster
from trying to wield a pick-axe if the square was reachable without digging,
was, while not exactly wrong, not the best solution.  The real problem is that
the attempt for the monster to wield a pick-axe was done unconditionally, so
the check for whether the square needs to be dug to be reached should be made
when trying to wield.
5) The curse part of that fix was partly correct and partly wrong.  The proper
thing to check for is mon->weapon_check, so that the monster does try to wield
the pick-axe once, and only doesn't even bother when it's the second and
successive times around.
6) The check for a monster in monmove.c wielding a weapon because it is near
you must take into account whether it is fleeing.  If it does not, and the
monster happens to be fleeing into a wall, it will repeatedly switch back
and forth between weapon and pick-axe.  (I verified that this is independent
of 4) and happened with or without it.)

Modified Files:
	doc/fixes33.1 src/dogmove.c src/monmove.c