From: michael.allison@mnr.gov.on.ca Subject: patch: new warning
This changes the way warning is conveyed. It is a compile time
option controlled by NEW_WARNING.  It makes warning a desireable
and useful thing, in contrast to the somewhat annoying and
intrusive way it is done currently.  This patch is being
distributed with NEW_WARNING not defined by default.

Instead of having warning place messages on the message window
periodically, it takes a different approach.  The "warning flash"
for monsters that meet the warning level and proximity criteria
are displayed continuously at the location of the monster that
you are being warned about.

The 0, 1,2,3,4, and 5 symbols are the default symbols, representing
each of the warning levels to maintain the ability on a monochrome
display. Warning 0 is never displayed however. You can change the
symbols if desired via WARNINGS= in your config file.
This config file entry would be equivalent to the defaults representing
warning levels 0 1 2 3 4 and 5:
WARNINGS=48 49 50 51 52 53

In short, the change makes you aware of the danger level of
the monsters in your vicinity, even though you can't see them
yet, but it doesn't reveal any information about what the
monsters are.

Objects that convey warning against a particular monster,
such as Sting act a little different.  They reveal the
presence of the target monster type anywhere on the
current level.

Notes:
- Warning level 0 (white) is not displayed.
- The six new tiles for the new glyphs have been implemented but
  I have not been able to test any of the tile related code changes
  (sys/share/tilemap.c) and there may be problems with them.
- all the changes are ifdef'd via NEW_WARNING

Modified Files:
	include/config.h include/decl.h include/display.h
	include/extern.h include/flag.h include/prop.h include/rm.h
	include/youprop.h src/apply.c src/artifact.c src/detect.c
	src/display.c src/do_wear.c src/dokick.c src/drawing.c
	src/explode.c src/files.c src/hack.c src/mon.c src/monmove.c
	src/options.c src/pager.c src/teleport.c win/share/other.txt
	win/share/tilemap.c win/tty/wintty.c