From: arromdee@rahul.net Subject: patch: new warning plus comments
This typo needs to be fixed in the warning patch.

Once it is fixed, the patch has a similar problem to an earlier warning
proposal did: it doesn't work well with sleeping monsters.  The problem can be
reproduced as follows:
1) Set SHOPTYPE='Z' to auto-create zoos in debug mode.
2) Start a game and wish for a ring of warning and a ring of stealth.  Put them
on.  Teleport to a deep level (22 or so).
3) Optionally magic-map the level.
4) Wander around until you get near the treasure zoo.  Note that not all of the
warned-about monsters show anything until you've actually entered and exited
the treasure zoo.  You may or may not have to actually see the monster first
before it gives warnings.

Also note that it behaves differently for monsters that are in range when you
put the ring on, and monsters that got into range when you moved.  I haven't
the faintest idea how to fix either of these.

Modified Files:
	src/display.c