From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: ghost name handling
     Another place where name handling for ghosts hadn't been updated
to reflect the change that made their names be stored the same as
ordinary monsters.  The result was redundant name display if you got
killed by one.

     There's something odd with the 'C' command.  It is allowed on
shopkeepers and temple priests, but the assigned name won't be shown
during play.  This suppresses it from the killer reason too, but I
think such naming should be disallowed, like is done for the Wizard,
if it's going to have no effect.

Modified Files:
	doc/fixes33.1 src/end.c