From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: fix for linux termcap
     I don't have the original report which described this problem,
but by deliberately altering the termcap entry I'm using I was able
to reproduce it (resulting in walls and such showing up as letters,
the typical symptom when something messes up the character set used
for DECgraphics).  The report had a suggested fix with it, but I
don't remember what that was.  This seems to be adequate, although
the tty termcap code is so tangled with conditionals that I'm not
sure about that.

Modified Files:
	win/tty/termcap.c win/tty/wintty.c