From: Pat Rankin <rankin@eql.caltech.edu> Subject: patch: ranger quest bit
     Attempt to address one of Mike's reservations about the Wumpus
oriented feedback in the middle of the ranger's quest:  the false
warning about bottomless pits.  All this does is give an extra message
when a pit is visited for the first time, about it having a bottom,
leaving the issue of whether there might be others which don't up to
the player's imagination.

Modified Files:
	src/trap.c