From: "Warwick Allison" <warwick@trolltech.com> Subject: PATCH: change warning tiles to coloured "?" marks
Since the warning marks are conceptually similar to the white "?" used
for the invisible unknown monster (behind boulder, etc.), it makes
sense to use the same shape for the NEW_WARNING tiles.  The attached
patch changes the solid coloured blocks to coloured questionmarks.
The colours are the same as the original colour blocks, except for
warnings 1 and 2 which needed to be changed for a clear colour
gradient.  The "?" could be made thicker if you think it's not clear
enough.

Playtesting this showed it was quite effective.

Modified Files:
	win/share/other.txt