Write your own epitaph
This patch implements the ability to write your own epitaph, which will
be engraved on your headstone and saved in the bones file. It is not
displayed in the visual tombstone right before the game exits, nor is it
saved in the logfile or high score list. It only shows up on the
headstone.

When you die, you will be asked "Do you want to write your own epitaph?"
right *before* DYWYPI?. If you respond with y, then you will be prompted
with "What do you want your epitaph to be?" Answering "I shouldn't have
eaten the cockatrice corpse." (with the period) will be
displayed thus when someone loads your bones file:

Something is engraved here on the headstone.--More--
You read: "Here lies rpresser. I shouldn't have eaten the cockatrice
corpse.".

Although the epitaph is written into the bones file, it does NOT break
bones compatibility. The epitaph was already being written into the
bones file as part of the grave object -- this just changes how the
epitaph string is written.

It gets called right before the grave is created; and as far as I can
tell, that only happens after the system has decided to write a bones
file. So this will tell you if a bones file was written.

Retrieved from http://bilious.homelinux.org/?170