Fixes the crash when cutting a long worm in half
This patch fixes bug C343-324 "Cutting a long worm in two will crash the
game if the cut takes the worm to 1 HP or if long worms had become
extinct."

The crash occurred when clone_mon() in makemon.c doesn't return a new
monster because of extinction or low hp. The original code didn't check
for this possibility.

Retrieved from http://bilious.homelinux.org/?332