Adds different messages to death reason, if you die while busy doing something else.
Normally when you die and are in the middle of a multiturn [in]action,
you get "while helpless" appended to the death reason. This patch adds
different reasons depending on why you're helpless, for example "while
disrobing" or "while sleeping off a magical draught".

Retrieved from http://bilious.homelinux.org/?12