This patch improves the prompt when NetHack asks the player to specify a location.
In vanilla NetHack you can let the cursor directly jump to stairs by
pressing '>' or '<', to fountains by '{', to altars by '_', etc.

But only if the location isn't covered by objects and for sinks and
grave this mechanism doesn't work as '|' and '#' also matches for
walls and corridors.

This is mainly an issue if your using the travel command '_' to quickly
navigate through the dungeon.

With this patch NetHack also finds dungeon features hidden by objects
and corridors and walls are never selected as target locations.

Bug reports can be directed to:
http://apps.sf.net/trac/unnethack/newticket

Retrieved from http://bilious.homelinux.org/?371