Replaces the PRNG used in NetHack with cryptographically secure algorithms.
This patch implements AES and SHA256 in rnd.c and uses them to generate
random numbers in a simpler Yarrow -like manner. It collects entropy
from /dev/random or /dev/urandom if LINUX is defined and high-resolution
clock (gettimeofday). If UNIX is not defined, it only uses the latter
for entropy (making this PRNG much weaker).

This patch is useful if you have a NetHack server and don't want your
players to crack the default PRNG (shown to be easily breakable).

Retrieved from http://bilious.homelinux.org/?349