Allows the user to change the colors of the dungeon features and traps, by setting options in the config file.
This patch works in all window ports that use character graphics, and
has been tested on the TTY, X11, and Qt ports.

Changing the colors can be done with putting the following lines into
the configuration file:
<pre>
DUNGEONCOLOR= 0 7 7 7 7 7 7 7 7 7 \
7 7 7 3 3 3 3 6 2 7 \
7 7 7 7 3 3 7 7 11 7 \
4 4 6 1 3 3 3 3 6 7 \
4

TRAPCOLORS= 6 6 7 3 6 1 7 12 4 9 0 0 3 3 5 5 13 7 7 12 12 10
</pre>
``DUNGEONCOLOR'' are in the same order as ``DUNGEON'', and
``TRAPCOLORS'' as ``TRAPS''

And here's another set of colors for the dungeon, with unlit
corridors darker than lit, and white stairs, just like Slash'em.
<pre>
DUNGEONCOLOR= 0 7 7 7 7 7 7 7 7 7 \
7 7 7 3 3 3 3 6 2 7 \
0 7 15 15 3 3 7 7 11 7 \
4 4 6 1 3 3 3 3 6 7 \
4
</pre>

The colors are:
0 black
1 red
2 green
3 brown
4 blue
5 magenta
6 cyan
7 gray
8 "no color"
9 orange
10 light green
11 yellow
12 light blue
13 light magenta
14 light cyan
15 white

Retrieved from http://bilious.homelinux.org/?17