Changes the option 'fixinv' to compound option, allowing finer degree of control.
Controls whether objects remember the inventory letters they had, and
how the inventory is reordered when picking up an object. The possible
values are:

none - objects do not remember their letters, and empty places in
inventory are compacted away;
next - objects remember their letters, but if the place is taken, put
the object to next free place.
move - objects remember their letters, but if the place is taken, move
the object underneath to next free place.

For backward compatibility, no value needs to be specified (which
defaults to 'next'), or it can be negated (which defaults to 'none').

Retrieved from http://bilious.homelinux.org/?22