Ring of Power Improvements
Hobbits have a 1 in 60 chance of spawning with the ring of power.
If you step into a fire trap while carrying the ring of fire, the
inscription on it will become temporarily readable. If you are a wizard,
minor, or major demon, it will be english. Otherwise, it will be
illegible.