NRH 3 map & RTP monster
NRH 3 map added, soon to have drink and big red button as well as other CSH specific things. RTP monster added. As of right now, it works like a guard in Minetown, but that will be fixed to be more appropriate soon. Also coming will be new dialog and sounds to better theme the level.