Correctly handle the SEGV achievement.
Previously, signal(3) was being used to set the signal handler. This
routine masks the signal to be delivered before delivering it. The
effect of this is that SEGVs generated within the SEGV handler will be
deferred until the SEGV handler completed. Of course, if the SEGV
handler is generating a SEGV, and the handler that would be invoked to
kill the program is being deferred, the program will hang forever.

This commit uses sigaction(2) instead to set the signal handler and uses
SA_RESETHAND | SA_NODEFER to reset the signal handler to default and
_not_ mask SEGV when a SEGV is delivered.