Print configfile error messages on failure.
If the configfile loading routine can't find (some) files, it suppresses
the error message. This was so that the user didn't see a flood of error
messages not being able to find config files in all the locations the
system looks for them. This change will prevent suppression of these
messages if the system can't find _any_ config files.

This should help with #111.